backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
python-dotenv==1.0.1
groq==1.2.0
httpx==0.28.1
python-multipart==0.0.9
bcrypt==4.2.1
PyJWT==2.9.0
redis==6.0.0
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.131.0
